+++ src/net/sourceforge/kolmafia/maximizer/CheckedItem.java
@@ -214,7 +214,7 @@ public class CheckedItem extends AdventureResult {
     }
 
     // Check mall price
-    int price = MallPriceManager.getMallPrice(this.getItemId());
+    long price = MallPriceManager.getMallPrice(this.getItemId());
 
     // Check if too expensive for max price settings
     if (price <= 0 || price > maxPrice) {
@@ -251,7 +251,7 @@ public class CheckedItem extends AdventureResult {
 
   public int inventory;
   public int initial;
-  public int creatable;
+  public long creatable;
   public int npcBuyable;
   public int mallBuyable;
   public int foldable;
